Moved config validator from domain to provider
