+++ dbt_platform_helper/providers/config.py
@@ -4,8 +4,8 @@ from pathlib import Path
 from schema import SchemaError
 
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
-from dbt_platform_helper.domain.config_validator import ConfigValidator
-from dbt_platform_helper.domain.config_validator import ConfigValidatorError
+from dbt_platform_helper.providers.config_validator import ConfigValidator
+from dbt_platform_helper.providers.config_validator import ConfigValidatorError
 from dbt_platform_helper.providers.io import ClickIOProvider
 from dbt_platform_helper.providers.platform_config_schema import PlatformConfigSchema
 from dbt_platform_helper.providers.yaml_file import FileNotFoundException
