+++ dbt_platform_helper/utils/validation.py
@@ -1,6 +1,6 @@
 from schema import SchemaError
 
-from dbt_platform_helper.domain.config_validator import ConfigValidator
+from dbt_platform_helper.providers.config_validator import ConfigValidator
 from dbt_platform_helper.providers.platform_config_schema import PlatformConfigSchema
